+++ tests/unit/devices/WH31.spec.ts
@@ -45,7 +45,7 @@ configs.forEach(config => {
 
       expect(device.battery.characteristics[0].value).to.equal(1); // low batt
       expect(device.humidity.service.characteristics[0].value).to.equal("Humidity 49 %")
-      expect(device.temperature.service.characteristics[0].value).to.equal("Temperature 80.78°F")
+      expect(device.temperature.service.characteristics[0].value).to.equal("Temperature 80.60°F")
       done();
     });